src/scripts/requirements.txt
websocket-client==1.6.1
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0